requirements.txt
streamlit==1.32.0
python-docx==1.0.0
python-dotenv==1.0.1
openai==1.12.0
PyPDF2==3.0.1
docx2txt==0.8
